+++ application/views/pages/recovery.php
@@ -7,19 +7,19 @@
         'assets/img/logo.png',
     ) ?>" alt="Easy!Appointments" class="shadow mb-3" width="72" height="72">
     <h4 class="text-primary fw-semibold mb-1"><?= lang('forgot_your_password') ?></h4>
-    <p class="text-secondary small mb-0"><?= lang('type_username_and_email_for_new_password') ?></p>
+    <p class=" small mb-0"><?= lang('type_username_and_email_for_new_password') ?></p>
 </div>
 
 <div class="alert d-none"></div>
 
 <form>
     <div class="mb-3">
-        <label for="username" class="form-label fw-medium text-secondary">
+        <label for="username" class="form-label fw-medium ">
             <?= lang('username') ?>
         </label>
         <div class="input-group">
             <span class="input-group-text bg-light border-end-0">
-                <i class="fas fa-user text-secondary"></i>
+                <i class="fas fa-user "></i>
             </span>
             <input type="text" id="username" 
                    placeholder="<?= lang('enter_username_here') ?>" class="form-control border-start-0 ps-0"/>
@@ -27,12 +27,12 @@
     </div>
 
     <div class="mb-4">
-        <label for="email" class="form-label fw-medium text-secondary">
+        <label for="email" class="form-label fw-medium ">
             <?= lang('email') ?>
         </label>
         <div class="input-group">
             <span class="input-group-text bg-light border-end-0">
-                <i class="fas fa-envelope text-secondary"></i>
+                <i class="fas fa-envelope "></i>
             </span>
             <input type="text" id="email" 
                    placeholder="<?= lang('enter_email_here') ?>" class="form-control border-start-0 ps-0"/>
@@ -47,7 +47,7 @@
     </div>
 
     <div class="text-center">
-        <a href="<?= site_url('login') ?>" class="text-decoration-none text-secondary small">
+        <a href="<?= site_url('login') ?>" class="text-decoration-none  small">
             <i class="fas fa-arrow-left me-1"></i>
             <?= lang('go_to_login') ?>
         </a>